+++ be/controllers/productController.ts
@@ -2,7 +2,6 @@ import { Request, Response } from "express";
 import cloudinary from "../utilities/Cloudinary";
 import Product from "../models/productModel";
 import Category from "../models/categoryModel";
-import test from "node:test";
 export const getProducts = async (req: Request, res: Response) => {
   try {
     const products = await Product.find().populate("categoryId");
@@ -26,6 +25,8 @@ export const createProduct = async (req: Request, res: Response) => {
       subCategoryName,
       brandName,
     } = JSON.parse(req.body.product);
+    const checkCoindence = await Product.findOne({ productCode: productCode });
+    if (checkCoindence) return res.status(403).send("Product code coincided");
     const images: any = req.files;
     const urlContainer: String[] = [];
     for (let i = 0; i < images?.length; i++) {
@@ -70,121 +71,86 @@ export const editProduct = async (req: Request, res: Response) => {
       `${new Date().getMonth() + 1}` +
       "/" +
       `${new Date().getDate()}`;
-    if (req.body.product.images?.length && req.files?.length) {
-      const productWithNewImages = JSON.parse(req.body.product);
-      const {
-        name,
-        description,
-        price,
-        productCode,
-        quantity,
-        tag,
-        disCount,
-        categoryName,
-        subCategoryName,
-        brandName,
-        _id,
-        images,
-      } = productWithNewImages;
-      const newImages: any = req.files;
-      let urlContainer: String[] = [];
-      for (let i = 0; i < newImages?.length; i++) {
-        const url: any = await uploadImg(newImages[i]);
-        urlContainer.push(url);
-      }
-      urlContainer = [...urlContainer, ...images];
-      const selectedCategory = await Category.findOne({ name: categoryName });
-      const product = await Product.findOneAndUpdate(
-        { _id: _id },
-        {
-          name,
-          description,
-          price,
-          productCode,
-          quantity,
-          tag,
-          disCount,
-          categoryId: selectedCategory?._id,
-          subCategoryName,
-          brandName,
-          images: urlContainer,
-          _id,
-          updatedAt: date,
-        }
-      );
-      return res.status(200).json({ msg: "Updated" });
-    }
     if (req.body.product) {
       const productWithNewImages = JSON.parse(req.body.product);
-      const {
-        name,
-        description,
-        price,
-        productCode,
-        quantity,
-        tag,
-        disCount,
-        categoryName,
-        subCategoryName,
-        brandName,
-        _id,
-      } = productWithNewImages;
-      const images: any = req.files;
-      const urlContainer: String[] = [];
-      for (let i = 0; i < images?.length; i++) {
-        const url: any = await uploadImg(images[i]);
-        urlContainer.push(url);
-      }
-      const selectedCategory = await Category.findOne({ name: categoryName });
-      const product = await Product.findOneAndUpdate(
-        { _id: _id },
-        {
-          name,
-          description,
-          price,
-          productCode,
-          quantity,
-          tag,
-          disCount,
-          categoryId: selectedCategory?._id,
-          subCategoryName,
-          brandName,
-          images: urlContainer,
-          _id,
-          updatedAt: date,
+      if (req.files?.length && productWithNewImages.images) {
+        const newImages: any = req.files;
+        let urlContainer: String[] = [];
+        for (let i = 0; i < newImages?.length; i++) {
+          const url: any = await uploadImg(newImages[i]);
+          urlContainer.push(url);
         }
-      );
-      return res.status(200).json({ msg: "Updated" });
+        urlContainer = [...urlContainer, ...productWithNewImages.images];
+        const selectedCategory = await Category.findOne({
+          name: productWithNewImages.categoryName,
+        });
+        const product = await Product.findOneAndUpdate(
+          { _id: productWithNewImages._id },
+          {
+            name: productWithNewImages.name,
+            description: productWithNewImages.description,
+            price: productWithNewImages.price,
+            productCode: productWithNewImages.productCode,
+            quantity: productWithNewImages.quantity,
+            tag: productWithNewImages.tag,
+            disCount: productWithNewImages.disCount,
+            categoryId: selectedCategory?._id,
+            subCategoryName: productWithNewImages.subCategoryName,
+            brandName: productWithNewImages.brandName,
+            images: urlContainer,
+            _id: productWithNewImages._id,
+            updatedAt: date,
+          }
+        );
+        return res.status(200).json({ msg: "Updated" });
+      } else {
+        const images: any = req.files;
+        const urlContainer: String[] = [];
+        for (let i = 0; i < images?.length; i++) {
+          const url: any = await uploadImg(images[i]);
+          urlContainer.push(url);
+        }
+        const selectedCategory = await Category.findOne({
+          name: productWithNewImages.categoryName,
+        });
+        const product = await Product.findOneAndUpdate(
+          { _id: productWithNewImages._id },
+          {
+            name: productWithNewImages.name,
+            description: productWithNewImages.description,
+            price: productWithNewImages.price,
+            productCode: productWithNewImages.productCode,
+            quantity: productWithNewImages.quantity,
+            tag: productWithNewImages.tag,
+            disCount: productWithNewImages.disCount,
+            categoryId: selectedCategory?._id,
+            subCategoryName: productWithNewImages.subCategoryName,
+            brandName: productWithNewImages.brandName,
+            images: urlContainer,
+            _id: productWithNewImages._id,
+            updatedAt: date,
+          }
+        );
+        return res.status(200).json({ msg: "Updated" });
+      }
     } else {
-      const {
-        name,
-        description,
-        price,
-        productCode,
-        quantity,
-        tag,
-        disCount,
-        categoryName,
-        subCategoryName,
-        brandName,
-        images,
-        _id,
-      } = req.body;
-      const selectedCategory = await Category.findOne({ name: categoryName });
+      const selectedCategory = await Category.findOne({
+        name: req.body.categoryName,
+      });
       const product = await Product.findByIdAndUpdate(
-        { _id: _id },
+        { _id: req.body._id },
         {
-          name,
-          description,
-          price,
-          productCode,
-          quantity,
-          tag,
-          disCount,
+          name: req.body.name,
+          description: req.body.description,
+          price: req.body.price,
+          productCode: req.body.productCode,
+          quantity: req.body.quantity,
+          tag: req.body.tag,
+          disCount: req.body.disCount,
           categoryId: selectedCategory?._id,
-          subCategoryName,
-          brandName,
-          images,
+          subCategoryName: req.body.subCategoryName,
+          brandName: req.body.brandName,
+          images: req.body.images,
           updatedAt: date,
         }
       );
